reformat functions to fetch updates and add logout to admin page
