+++ client/src/files/data/recordingsData.js
@@ -1,3 +1,3 @@
-const recordingsData = [{"date":"May 14, 2023","anthem":"I'll Be On My Way","anthem-type":"audio","anthem-src":"/audio/I'll Be On My Way.mp3","offertory":"The Lord Bless You and Keep You","offertory-type":"audio","offertory-src":"/audio/The Lord Bless You and Keep You.mp3"},{"date":"April 30, 2023","anthem":"Modimo","anthem-type":"audio","anthem-src":"/audio/Modimo.mp3","offertory":"The Time is Near / I Will March","offertory-type":"audio","offertory-src":"/audio/The Time is Near _ I Will March.mp3"},{"date":"April 9, 2023","anthem":"Hallelujah from the Mount of Olives","anthem-type":"audio","anthem-src":"/audio/Hallelujah from the Mount of Olives.mp3","offertory":"Hallelujah Chorus","offertory-type":"audio","offertory-src":"/audio/LGPC Hallelujah Chorus Orchestra + Organ NO METRONOME v10_mixdown.mp3"},{"date":"April 23, 2023","anthem":"How Can I Keep from Singing","anthem-type":"audio","anthem-src":"/audio/8 - How Can I Keep From Singing.mp3","offertory":"All Earth is Hopeful","offertory-type":"audio","offertory-src":"/audio/16 - All Earth is Hopeful.mp3"},{"date":"April 9, 2023","anthem":"Hallelujah from the Mount of Olives","anthem-type":"youtube","anthem-src":"https://www.youtube-nocookie.com/embed/xK-IHpN1HBs","offertory":"Hallelujah Chorus","offertory-type":"youtube","offertory-src":"https://www.youtube-nocookie.com/embed/BBZ7AfZR9xs"},{"date":"April 23, 2023","anthem":"How Can I Keep from Singing","anthem-type":"youtube","anthem-src":"https://www.youtube-nocookie.com/embed/rL3-gqGHfdw","offertory":"All Earth is Hopeful","offertory-type":"youtube","offertory-src":"https://www.youtube-nocookie.com/embed/dq5BEfErVVY"}]
+const recordingsData = [{"date":"April 9, 2023","anthem":"Hallelujah from the Mount of Olives","anthem-type":"audio","anthem-src":"/audio/Hallelujah from the Mount of Olives.mp3","offertory":"Hallelujah Chorus","offertory-type":"audio","offertory-src":"/audio/LGPC Hallelujah Chorus Orchestra + Organ NO METRONOME v10_mixdown.mp3"},{"date":"April 23, 2023","anthem":"How Can I Keep from Singing","anthem-type":"audio","anthem-src":"/audio/8 - How Can I Keep From Singing.mp3","offertory":"All Earth is Hopeful","offertory-type":"audio","offertory-src":"/audio/16 - All Earth is Hopeful.mp3"},{"date":"April 30, 2023","anthem":"Modimo","anthem-type":"audio","anthem-src":"/audio/Modimo.mp3","offertory":"The Time is Near / I Will March","offertory-type":"audio","offertory-src":"/audio/The Time is Near _ I Will March.mp3"},{"date":"May 14, 2023","anthem":"I'll Be On My Way","anthem-type":"audio","anthem-src":"/audio/I'll Be On My Way.mp3","offertory":"The Lord Bless You and Keep You","offertory-type":"audio","offertory-src":"/audio/The Lord Bless You and Keep You.mp3"},{"date":"April 9, 2023","anthem":"Hallelujah from the Mount of Olives","anthem-type":"youtube","anthem-src":"https://www.youtube-nocookie.com/embed/xK-IHpN1HBs","offertory":"Hallelujah Chorus","offertory-type":"youtube","offertory-src":"https://www.youtube-nocookie.com/embed/BBZ7AfZR9xs"},{"date":"April 23, 2023","anthem":"How Can I Keep from Singing","anthem-type":"youtube","anthem-src":"https://www.youtube-nocookie.com/embed/rL3-gqGHfdw","offertory":"All Earth is Hopeful","offertory-type":"youtube","offertory-src":"https://www.youtube-nocookie.com/embed/dq5BEfErVVY"}]
 
 export { recordingsData }
