+++ server/views/admin-login.html
@@ -19,15 +19,11 @@
     </form>
   </main>
 
-
-
 <style>
   body {
   display: flex;
   justify-content: center;
   align-items: center;
-  background-color: black;
-  color: white;
   text-align: center;
   }
 </style>
