+++ server/views/admin.html
@@ -5,13 +5,26 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Choir Admin</title>
-  <link rel="stylesheet" href="./style.css">
 </head>
 <body>
   <main>
     <h1>This is the Admin Dashboard</h1>
+    <p>Update Google sheet <a href="https://docs.google.com/spreadsheets/d/1whUSpnEkLS0xtPdvKd0Z27Lm2T3sQMQphhTkh-GAXq8/edit?usp=sharing" target="_blank">here</a>, then submit the changes below.</p>
+    <form action="" method="post">
+      <label for="makeUpdate">Update 'schedule' or 'recordings', or 'logout'. </label>
+      <input type="text" name="makeUpdate">
+      <button type="submit">Submit</button>
+    </form>
   </main>
 
+<style>
+  body {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  }
+</style>
